backend/requirements.txt
# AWS Lambda用掲示板アプリケーション依存パッケージ
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.2
boto3==1.33.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
moto==4.2.11
httpx==0.25.2
python-multipart==0.0.6
